requirements.txt
Flask==1.1.2
gevent==20.6.1
pandas==1.0.4
Pillow==7.1.2
sklearn